fix: address review comment
