+++ src/elements/content-sidebar/__tests__/MetadataSidebarRedesign.test.tsx
@@ -11,6 +11,22 @@ import {
 } from '../MetadataSidebarRedesign';
 import useSidebarMetadataFetcher, { STATUS } from '../hooks/useSidebarMetadataFetcher';
 import useMetadataFieldSelection from '../hooks/useMetadataFieldSelection';
+import type { MetadataInstanceEditorProps } from '../MetadataInstanceEditor';
+
+// Records props passed to the editor so wiring tests can assert on them
+// without replacing the real editor used by the rest of this suite.
+const mockEditor = jest.fn<null, [MetadataInstanceEditorProps]>(() => null);
+jest.mock('../MetadataInstanceEditor', () => {
+    const actual = jest.requireActual('../MetadataInstanceEditor') as typeof import('../MetadataInstanceEditor');
+
+    return {
+        __esModule: true,
+        default: (props: MetadataInstanceEditorProps) => {
+            mockEditor(props);
+            return actual.default(props);
+        },
+    };
+});
 
 jest.mock('../hooks/useSidebarMetadataFetcher');
 const mockUseSidebarMetadataFetcher = useSidebarMetadataFetcher as jest.MockedFunction<
@@ -653,6 +669,35 @@ describe('elements/content-sidebar/Metadata/MetadataSidebarRedesign', () => {
         renderComponent({ getPreview });
 
         expect(mockUseMetadataFieldSelection).toHaveBeenCalledWith(getPreview);
+    });
+
+    test('passes host-provided fetchers to the editor when the user field flag is on', async () => {
+        const fetchUsers = jest.fn();
+        const fetchAvatarUrls = jest.fn();
+
+        mockUseSidebarMetadataFetcher.mockReturnValue({
+            clearExtractError: jest.fn(),
+            extractSuggestions: jest.fn(),
+            handleCreateMetadataInstance: jest.fn(),
+            handleDeleteMetadataInstance: jest.fn(),
+            handleUpdateMetadataInstance: jest.fn(),
+            templateInstances: [mockVisibleTemplateInstance],
+            templates: mockTemplates,
+            errorMessage: null,
+            status: STATUS.SUCCESS,
+            file: mockFile,
+            extractErrorCode: null,
+        });
+
+        renderComponent({ fetchAvatarUrls, fetchUsers }, { 'metadata.userField.enabled': true });
+
+        await userEvent.click(screen.getByRole('button', { name: 'Edit Visible Template' }));
+
+        const { calls } = mockEditor.mock;
+        const [editorProps] = calls[calls.length - 1];
+        expect(editorProps.isMetadataUserFieldEnabled).toBe(true);
+        expect(editorProps.fetchUsers).toBe(fetchUsers);
+        expect(editorProps.fetchAvatarUrls).toBe(fetchAvatarUrls);
     });
 
     describe('navigation blocking', () => {
